refactor: 공통으로 사용하는 domain, exception은 common 패키지로 이동
