+++ src/main/java/roomescape/global/common/exception/AlreadyExistsException.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception;
+package roomescape.global.common.exception;
 
 public class AlreadyExistsException extends RuntimeException {
