+++ src/main/java/roomescape/global/common/exception/DuplicatedException.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception;
+package roomescape.global.common.exception;
 
 public class DuplicatedException extends RuntimeException {
