+++ src/main/java/roomescape/reservation/exception/ReservationAlreadyExistsException.java
@@ -1,6 +1,6 @@
 package roomescape.reservation.exception;
 
-import roomescape.global.exception.AlreadyExistsException;
+import roomescape.global.common.exception.AlreadyExistsException;
 
 public class ReservationAlreadyExistsException extends AlreadyExistsException {
