+++ src/main/java/roomescape/reservationtime/exception/ReservationTimeAlreadyExistsException.java
@@ -1,6 +1,6 @@
 package roomescape.reservationtime.exception;
 
-import roomescape.global.exception.AlreadyExistsException;
+import roomescape.global.common.exception.AlreadyExistsException;
 
 public class ReservationTimeAlreadyExistsException extends AlreadyExistsException {
